Fix bug : EXTRA_ELEMENTS ~ 4 OK
